Moved refactor to CinemachineUpgradeManager and removed temporary refactor file that stored the main refactor steps.
