--- com.unity.cinemachine/Editor/Upgrader/CinemachineUpgradeManager.refactor.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: fa29d08d426c458c8264cbd84e4535ec
-timeCreated: 1662557666
